refactor: extract README generator into src/markdown/repo-readme.js (#32)
Move generateReadmeFromOutline and scoreBadgeMd out of score-report.js into a
dedicated src/markdown/repo-readme.js module. score-report.js re-exports
scoreBadgeMd so the code-quality badge stays shared with SCORE.md consumers.
Importers (repo-apply, apply-all, test-pipeline, tests) now pull the README
generator from the new module. No behavior change; existing apply flow pushes
identical output.
