+++ api/repo-apply.js
@@ -1,7 +1,8 @@
 import { getRepoSnapshot }                        from '../src/github/repo-contents.js';
 import { analyzeRepo }                            from '../src/ai/repo-analyzer.js';
 import { scanCache }                              from '../src/scan-cache.js';
-import { generateScoreReport, generateReadmeFromOutline, scoreBadgeMd } from '../src/markdown/score-report.js';
+import { generateScoreReport }                       from '../src/markdown/score-report.js';
+import { generateReadmeFromOutline, scoreBadgeMd }   from '../src/markdown/repo-readme.js';
 
 const ghHeaders = (token) => ({
     Authorization:  `Bearer ${token}`,
