+++ src/markdown/score-report.js
@@ -1,22 +1,9 @@
+// The code-quality badge lives with the README generator; re-exported here so
+// SCORE.md consumers can keep importing it from this module unchanged.
+export { scoreBadgeMd } from './repo-readme.js';
+
 const bar  = (score) => '█'.repeat(Math.round(score / 10)) + '░'.repeat(10 - Math.round(score / 10));
 const grade = (g) => ({ 'A+': '🟢', A: '🟢', B: '🔵', C: '🟡', D: '🟠', F: '🔴' }[g] ?? '⚪');
-
-const BADGE_COLORS = { 'A+': 'brightgreen', A: 'green', B: 'blue', C: 'yellow', D: 'orange', F: 'red' };
-
-/**
- * Returns the shields.io badge markdown for a repo's current code quality score.
- * Intended to be embedded in README.md and updated each time SCORE.md is regenerated.
- *
- * @param {object} analysis  Result of analyzeRepo()
- */
-export const scoreBadgeMd = (analysis) => {
-    const { overall, grade: g } = analysis.codeQuality;
-    const color = BADGE_COLORS[g] ?? 'lightgrey';
-    // shields.io path encoding: _ = space, %2B = +, %2F = /, %C2%B7 = ·
-    const msg = `${g}_%C2%B7_${overall}%2F100`.replace('+', '%2B');
-    const url = `https://img.shields.io/badge/code_quality-${msg}-${color}?style=flat-square`;
-    return `[![Code Quality](${url})](SCORE.md)`;
-};
 
 const dimSection = (label, dim) => {
     if (!dim) return '';
@@ -86,63 +73,6 @@ export const generateScoreReport = (repoName, analysis) => {
 
     parts.push('---');
     parts.push('_Generated by [github-pretty-readme](https://github.com/kevinthelago/github-pretty-readme)_');
-
-    return parts.join('\n');
-};
-
-/**
- * Generates a README.md from a scan's readmeOutline.
- *
- * @param {string} repoName
- * @param {object} analysis  Result of analyzeRepo()
- */
-export const generateReadmeFromOutline = (repoName, analysis) => {
-    const { readmeOutline, techStack, meta } = analysis;
-    if (!readmeOutline) return null;
-
-    const { title, tagline, features, installationSteps, usageExample } = readmeOutline;
-    const parts = [];
-
-    parts.push(`# ${title || repoName}`, '');
-    if (tagline) parts.push(`> ${tagline}`, '');
-
-    if (meta?.language) {
-        parts.push(`![${meta.language}](https://img.shields.io/badge/-${encodeURIComponent(meta.language)}-555?style=flat)`);
-    }
-    if (meta?.license) {
-        parts.push(`![License](https://img.shields.io/badge/license-${encodeURIComponent(meta.license)}-blue?style=flat)`);
-    }
-    if (meta?.language || meta?.license) parts.push('');
-
-    if (features?.length) {
-        parts.push('## Features', '');
-        features.forEach(f => parts.push(`- ${f}`));
-        parts.push('');
-    }
-
-    if (installationSteps?.length) {
-        parts.push('## Installation', '');
-        parts.push('```bash');
-        installationSteps.forEach(s => parts.push(s));
-        parts.push('```', '');
-    }
-
-    if (usageExample) {
-        parts.push('## Usage', '');
-        parts.push('```');
-        parts.push(usageExample);
-        parts.push('```', '');
-    }
-
-    if (techStack?.length) {
-        parts.push('## Built With', '');
-        parts.push(techStack.map(t => `- ${t}`).join('\n'), '');
-    }
-
-    if (meta?.license) {
-        parts.push('## License', '');
-        parts.push(`[${meta.license}](LICENSE)`, '');
-    }
 
     return parts.join('\n');
 };
